+++ Random.h
@@ -6,7 +6,7 @@ class Random
 public:
 	Random() = delete;
 	static int getRandomNumber(int rangeStart, int rangeEnd);
-	static bool isSuccessful(float probability);
+	static bool isSuccessful(double probability);
 private:
 	static std::random_device rd;
 	static std::mt19937 mt;
